Inline start_server fn into payas_server main
This is the only place the function is used now (since there is no
longer a dev/prod mode separation).
